requirements.txt
watchdog
requests
pystray
pillow